Add footer example project
