+++ JoyfillSwiftUIExample/JoyfillExample/UserAccessTokenTextFieldView.swift
@@ -480,6 +480,7 @@ struct OptionSelectionView: View {
         case metadataChangeAPIDemo
         case simpleForm
         case simpleNavigationTest
+        case footerExample
 
         var title: String {
             switch self {
@@ -513,9 +514,11 @@ struct OptionSelectionView: View {
                 return "Simple example Form"
             case .simpleNavigationTest:
                 return "Navigation Test"
+            case .footerExample:
+                return "Footer Example"
             }
         }
-        
+
         var description: String {
             switch self {
             case .token:
@@ -548,9 +551,11 @@ struct OptionSelectionView: View {
                 return "Simple example Form"
             case .simpleNavigationTest:
                 return "Navigate to any page or field in the form"
+            case .footerExample:
+                return "Test footer show/hide, expand, and keyboard behaviour"
             }
         }
-        
+
         var icon: String {
             switch self {
             case .token:
@@ -583,9 +588,11 @@ struct OptionSelectionView: View {
                 return "slider.horizontal.3"
             case .simpleNavigationTest:
                 return "location.circle.fill"
+            case .footerExample:
+                return "dock.rectangle"
             }
         }
-        
+
         var color: Color {
             switch self {
             case .token:
@@ -618,6 +625,8 @@ struct OptionSelectionView: View {
                 return .blue
             case .simpleNavigationTest:
                 return .indigo
+            case .footerExample:
+                return .mint
             }
         }
     }
@@ -765,6 +774,8 @@ struct OptionSelectionView: View {
             SimpleFormExampleView()
         case .simpleNavigationTest:
             SimpleNavigationTestView(showAlert: $showNavigationAlert, alertMessage: $navigationAlertMessage)
+        case .footerExample:
+            FooterExampleView()
         case .none:
             AnyView(EmptyView())
         case .some(.allFormulaJSONs):
